pyproject.toml
[project]
name = "tasks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.116.1",
    "gunicorn>=23.0.0",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.43",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "httpx>=0.28.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
